update profile command!

+++ src/commands/profile.sh
@@ -1,4 +1,8 @@
 profile() {
+    if ! [ -f "$HOME/.config/bash/profiles/$1.bash" ]; then
+        echo "Can't find profile!"
+        return
+    fi
     config=$(cat $HOME/.config/bash/custom/config.bdb)
     echo "PROFILE=$1;OSICON=$(bdbFind OSICON $config)" > $HOME/.config/bash/custom/config.bdb
     bash
